#2826 Refactored code into separate class
