+++ src/MCT.js
@@ -266,6 +266,7 @@ define([
         this.install(this.plugins.WebPage());
         this.install(this.plugins.Condition());
         this.install(this.plugins.ConditionWidget());
+        this.install(this.plugins.URLTimeSettingsSynchronizer());
     }
 
     MCT.prototype = Object.create(EventEmitter.prototype);
